Refactored and renamed game objects
